requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0
matplotlib>=3.3.0
diffusers>=0.16.0
transformers>=4.30.0
tqdm>=4.64.0
gradio>=3.50.0
imageio>=2.28.0
discord.py>=2.3.0
python-dotenv>=1.0.0 